+++ packages/angular/src/classes/observable-channel.class.spec.ts
