+++ packages/angular/src/classes/observable-channel.class.ts
