+++ packages/angular/src/interfaces/xmess-module.interface.ts
