+++ packages/core/src/classes/channel-tree.class.spec.ts
