+++ packages/core/src/classes/channel-tree.class.ts
