+++ packages/core/src/classes/channel.class.spec.ts
