+++ packages/core/src/classes/channel.class.ts
