+++ packages/core/src/classes/hook.class.spec.ts
