+++ packages/core/src/classes/hook.class.ts
