+++ packages/core/src/interfaces/base-object.interface.ts
