+++ packages/core/src/interfaces/channel-tree.interface.ts
