+++ packages/core/src/interfaces/channel.interface.ts
