+++ packages/core/src/interfaces/hook.interface.ts
