+++ packages/core/src/interfaces/path.interface.ts
