+++ packages/core/src/interfaces/xmess.interface.ts
