+++ packages/core/src/utils/path.util.spec.ts
