+++ packages/core/src/utils/path.util.ts
