+++ packages/global-context/src/classes/context.class.ts
