+++ packages/global-context/src/constants/global-context.constant.ts
