+++ packages/global-context/src/global-context.plugin.spec.ts
