+++ packages/global-context/src/global-context.plugin.ts
