+++ packages/global-context/src/interfaces/context.interface.ts
